requirements.txt
pandas
python-dotenv
pymongo
scikit-learn

-e .